fix: extract useUnreadNotifications hook, remove duplicate polling
- Create shared useUnreadNotifications hook with try/catch error handling
- Remove inline notification polling from Sidebar, BottomNav, MobileHeader
- Fix brittle badge string comparison (item.name) to use icon identity
- Revert unrelated formatting-only changes in 4 files
