+++ apps/web/src/components/shared/Sidebar.tsx
@@ -31,7 +31,7 @@ import { HouseIcon } from '@/components/shared/icons/HouseIcon';
 import { useAuth } from '@/hooks/useAuth';
 import { usePostHog } from '@/hooks/usePostHog';
 import { useUnreadMessages } from '@/hooks/useUnreadMessages';
-import { getAuthToken } from '@/lib/auth';
+import { useUnreadNotifications } from '@/hooks/useUnreadNotifications';
 import { getUserDisplayName } from '@/lib/user-display';
 
 /**
@@ -47,7 +47,6 @@ function SidebarContent() {
   const [collapsed, setCollapsed] = useState(false);
   const [showMdMenu, setShowMdMenu] = useState(false);
   const [copiedReferral, setCopiedReferral] = useState(false);
-  const [unreadNotifications, setUnreadNotifications] = useState(0);
   const [feedbackModalOpen, setFeedbackModalOpen] = useState(false);
   const asideRef = useRef<HTMLElement>(null);
   const mdMenuRef = useRef<HTMLDivElement>(null);
@@ -55,6 +54,7 @@ function SidebarContent() {
   const { ready, authenticated, user, logout, login } = useAuth();
   const { trackNavigation, trackClick } = usePostHog();
   const { totalUnread: unreadMessages } = useUnreadMessages();
+  const { unreadCount: unreadNotifications } = useUnreadNotifications();
 
   // Hide sidebar when WAITLIST_MODE is enabled on home page
   const isWaitlistMode = process.env.NEXT_PUBLIC_WAITLIST_MODE === 'true';
@@ -83,42 +83,6 @@ function SidebarContent() {
     }
     return undefined;
   }, [showMdMenu]);
-
-  // Poll for unread notifications
-  useEffect(() => {
-    if (!authenticated || !user) {
-      setUnreadNotifications(0);
-      return;
-    }
-
-    const fetchUnreadCount = async () => {
-      const token = getAuthToken();
-
-      if (!token) {
-        return;
-      }
-
-      const response = await fetch(
-        '/api/notifications?unreadOnly=true&limit=1',
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-
-      if (response.ok) {
-        const data = await response.json();
-        setUnreadNotifications(data.unreadCount || 0);
-      }
-    };
-
-    fetchUnreadCount();
-
-    // Refresh every 1 minute
-    const interval = setInterval(fetchUnreadCount, 60000); // 60 seconds = 1 minute
-    return () => clearInterval(interval);
-  }, [authenticated, user]);
 
   // Adjust sidebar height to account for elements above it (e.g. NFT banner)
   // so the user profile bar at the bottom is always visible
@@ -291,8 +255,8 @@ function SidebarContent() {
           {navItems.map((item) => {
             const Icon = item.icon;
             const hasNotificationBadge =
-              (item.name === 'Notifications' && unreadNotifications > 0) ||
-              (item.name === 'Chats' && unreadMessages > 0);
+              (Icon === Bell && unreadNotifications > 0) ||
+              (Icon === MessageCircle && unreadMessages > 0);
 
             const navContent = (
               <>
